requirements.txt
transformers
torch
torchvision
pillow
requests
